Adds tests for the parser
### Notes
This isn't by any means exhaustive testing, but it covers quite a few
cases in the parser.

Next steps: Add tests for the operation handler

Note: Quoted strings are not tested at all in this. I'm not sure if
that's something that should even be supported and may be removed prior
to sending out.

### Testing
* Ran unit tests
* Built and ran executable